requirements.txt
fastapi==0.118.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
celery==5.3.4
redis==5.0.1
alembic==1.13.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
moviepy==1.0.3
Pillow==9.5.0
email-validator==2.1.0
pytest==7.4.0
pytest-cov
pytest-asyncio
factory_boy
faker
httpx